separate out PatchCcoWithAppSwitchEligibilityResponse
