+++ Sources/CorePayments/PatchCcoWithAppSwitchEligibilityVariables.swift
@@ -1,7 +1,5 @@
 import Foundation
 
-// MARK: - Variables
-
 struct ExperimentationContext: Encodable {
 
     let integrationChannel: String?
@@ -22,28 +20,3 @@ struct PatchCcoWithAppSwitchEligibilityVariables: Encodable {
     let paypalNativeAppInstalled: Bool?
 }
 
-// MARK: - Response (data-only)
-
-@_documentation(visibility: private)
-public struct PatchCcoWithAppSwitchEligibilityResponse: Decodable {
-
-    let external: ExternalNode?
-
-    struct ExternalNode: Decodable {
-
-        let patchCcoWithAppSwitchEligibility: PatchNode?
-    }
-
-    struct PatchNode: Decodable {
-
-        let appSwitchEligibility: AppSwitchEligibility?
-    }
-}
-
-@_documentation(visibility: private)
-public struct AppSwitchEligibility: Decodable {
-    
-    public let appSwitchEligible: Bool?
-    public let redirectURL: String?
-    public let ineligibleReason: String?
-}
